ir: types.go -> node_types.go + move ifaces to ir.go (#738)
